libft: ft_striteri

+++ ft_strmapi.c
@@ -2,9 +2,9 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char	*str;
-	size_t	i;
-	size_t	len;
+	char			*str;
+	unsigned int	i;
+	size_t			len;
 
 	if (!s || !f)
 		return (NULL);
